wcnss: Fix buffer overflow in wcnss_prealloc_get
There is potential integer truncation in the wcnss_prealloc_get api.
size_t is 8 byte on x64 platform and "unsigned int" is 4 byte.
To avoid this integer truncation, pass size as size_t instead
of unsigned int.

CRs-Fixed: 2269610

+++ drivers/net/wireless/cnss_prealloc/cnss_prealloc.c
@@ -1,4 +1,4 @@
-/* Copyright (c) 2012,2014-2016 The Linux Foundation. All rights reserved.
+/* Copyright (c) 2012,2014-2016, 2018 The Linux Foundation. All rights reserved.
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License version 2 and
@@ -34,7 +34,7 @@ static struct dentry *debug_base;
 
 struct wcnss_prealloc {
 	int occupied;
-	unsigned int size;
+	size_t size;
 	void *ptr;
 #ifdef CONFIG_SLUB_DEBUG
 	unsigned long stack_trace[WCNSS_MAX_STACK_TRACE];
@@ -212,7 +212,7 @@ static inline void wcnss_prealloc_save_stack_trace(struct wcnss_prealloc *entry)
 }
 #endif
 
-void *wcnss_prealloc_get(unsigned int size)
+void *wcnss_prealloc_get(size_t size)
 {
 	int i = 0;
 	unsigned long flags;
@@ -232,8 +232,8 @@ void *wcnss_prealloc_get(unsigned int size)
 	}
 	spin_unlock_irqrestore(&alloc_lock, flags);
 
-	pr_err("wcnss: %s: prealloc not available for size: %d\n",
-			__func__, size);
+	pr_err("wcnss: %s: prealloc not available for size: %zu\n",
+	       __func__, size);
 
 	return NULL;
 }
@@ -259,7 +259,7 @@ int wcnss_prealloc_put(void *ptr)
 EXPORT_SYMBOL(wcnss_prealloc_put);
 
 #ifdef CONFIG_WCNSS_SKB_PRE_ALLOC
-struct sk_buff *wcnss_skb_prealloc_get(unsigned int size)
+struct sk_buff *wcnss_skb_prealloc_get(size_t size)
 {
 	int i = 0;
 	unsigned long flags;
@@ -279,7 +279,7 @@ struct sk_buff *wcnss_skb_prealloc_get(unsigned int size)
 	}
 	spin_unlock_irqrestore(&alloc_lock, flags);
 
-	pr_err("wcnss: %s: prealloc not available for size: %d\n",
+	pr_err("wcnss: %s: prealloc not available for size: %zu\n",
 	       __func__, size);
 
 	return NULL;
@@ -321,8 +321,8 @@ void wcnss_prealloc_check_memory_leak(void)
 			leak_detected = true;
 		}
 
-		pr_err("Size: %u, addr: %pK, backtrace:\n",
-				wcnss_allocs[i].size, wcnss_allocs[i].ptr);
+		pr_err("Size: %zu, addr: %pK, backtrace:\n",
+		       wcnss_allocs[i].size, wcnss_allocs[i].ptr);
 		print_stack_trace(&wcnss_allocs[i].trace, 1);
 	}
 
@@ -345,9 +345,8 @@ void wcnss_skb_prealloc_check_memory_leak(void)
 			leak_detected = true;
 		}
 
-		pr_err(
-			"Size: %u, addr: %pK, backtrace:\n",
-			wcnss_skb_allocs[i].size, wcnss_skb_allocs[i].ptr);
+		pr_err("Size: %zu, addr: %pK, backtrace:\n",
+		       wcnss_skb_allocs[i].size, wcnss_skb_allocs[i].ptr);
 		print_stack_trace(&wcnss_skb_allocs[i].trace, 1);
 	}
 }
